Rework `Project` and `SourceFile` lifecycles
They can now all be initialized, analyzed and refreshed with the least work possible. In the future we ideally only have to refresh files that actually changed and not all of them. The API for that is now available. This is probably something the consumer has to do manage themselves